refactor: Move getSubtitle function to utils and import to articleArchive and relatedArticles MIM-2058
